Added App Code
Created a TestApp directory with subdirectories for "src", "executable",
and "docs".  The compiled (.apk) file is in "executable" and the App
Inventor source file (.aia) is in "src".  A user document will be added
to "docs" later.  The code has been compiled for e-mail or USB loading
to an android device and not for the Google Play Store compatibility.
The app has been tested and appears to work well.  It does not process
the toggle switch data, however.