Fix for Indicator randomly turning off
I found the bug and this firmware version (1.1) fixes it.  The bug was
in loop() circa line 259.  There is a test for not being connected to
the Particle cloud and if it is true (not connected), the indicator is
flashed.  This test can end anytime the device reconnects to the cloud
and can leave the indicator in the off state because there is no code to
retest the pushbutton state until the latter changes.  I added an "else"
in line 261 which tests the current status of the pushbutton and sets
the indicator accordingly.